refactor(primer): rename primer module to primer_designer
